Started Prep for GMM Plant Segmentation
