+++ dev/spect_dev_1.py
@@ -1,7 +1,7 @@
 from src.lib import data
 
 
-Spect = data.Data()
+Spect = data.Data(dataset_day=1, scan_number=2)
 Spect.import_spect()
 Spect.histogram_analysis()
 Spect.noise_removal()
