+++ src/config/config.py
@@ -1,5 +1,5 @@
 config = {
 'data_path': './data/',
-'datasets': ['07Mar2018'],
+'datasets': ['07Mar2018', '12Mar2018'],
 'scans': ['CDT_GRP3_Ga67_PLANTS9(L)7(R)', 'CDT_GRP3_Ga67_PLANTS3(R)4(L)', 'CDT_GRP3_Ga67_PLANTS 1(L)2(R)']
 }
