+++ src/lib/Segmentation.py
@@ -1,0 +1,3 @@
+class PlantSegmentation:
+    def __init__(self, Spect):
+        euc_distro = Spect.main_comp
